Break out api specific stuff in datasource to its own file
